Update Some_CPT_Simu.Rmd
I have added code that will produce ARMA error. We could simplify this (if we wanted) as the arima.sim function can (i) simulate IID Gaussian or student-t; and (ii) simulate ARIMA models with non-Gaussian distributions. I.e. we have a set of "noise-models" we loop over which are defined in terms of the base distribution, any parameters of this, and the arima model.